// Hospital Project With API/hospitalng/src/app/shared/models/doctorDto.ts
export interface DoctorDto {
    doctorId: number; 
    firstName: string;
    lastName: string;
    username: string;
    password: string;
    country: string;
    email: string;
    phone: number;
    address: string;
    speciality: string;
    availability: string;
    hospital: string;
    department: string;
  }